Final Upload With Repeat Changes
This commit contains several changes made during the repeat process of the project. Including general touch-ups to code, as well as the implementation of an approval system for the corridors; fulfilling all basic requirements for the project. Documents have been updated with an additional sections pertaining to what I did during the repeat. User Manual fully redone